Microoptimizations
